update registro y login, ajustes permisos

+++ ayuda_municipios/src/main/java/com/example/ayuda_municipios/municipios/MunicipiosService.java
@@ -3,12 +3,16 @@ package com.example.ayuda_municipios.municipios;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.example.ayuda_municipios.municipios.dto.MunicipioDTO;
 import com.example.ayuda_municipios.provincias.Provincia;
 import com.example.ayuda_municipios.provincias.ProvinciasRepository;
+import com.example.ayuda_municipios.usuarios.Usuario;
+import com.example.ayuda_municipios.usuarios.UsuariosRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,6 +21,8 @@ import lombok.RequiredArgsConstructor;
 public class MunicipiosService {
     private final MunicipiosRepository municipiosRespository;
     private final ProvinciasRepository provinciasRespository;
+    private final UsuariosRepository usuariosRepository;
+
 
     public List<Municipio> verMunicipios() {
         return municipiosRespository.findBy();
@@ -50,6 +56,15 @@ public class MunicipiosService {
     }
 
     public void delete(int id) {
-        municipiosRespository.deleteById(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Integer idAuth = Integer.parseInt(auth.getCredentials().toString());
+        Usuario usuarioActual= usuariosRepository.findUsuarioById(idAuth);
+
+        if (usuarioActual.getRol()=="ADMIN") {
+            municipiosRespository.deleteById(id);
+        }else{
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Solo los administradores tienen permisos para eliminar");
+        }
+
     }
 }
