+++ ayuda_municipios/src/main/java/com/example/ayuda_municipios/provincias/ProvinciasService.java
@@ -3,10 +3,14 @@ package com.example.ayuda_municipios.provincias;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.example.ayuda_municipios.provincias.dto.ProvinciaDTO;
+import com.example.ayuda_municipios.usuarios.Usuario;
+import com.example.ayuda_municipios.usuarios.UsuariosRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -14,6 +18,8 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ProvinciasService {
     private final ProvinciasRepository provinciasRepository;
+        private final UsuariosRepository usuariosRepository;
+
 
     public List<Provincia> verProvincias(){
         return provinciasRepository.findBy();
@@ -45,6 +51,15 @@ public class ProvinciasService {
     }
 
     public void delete(int id) {
-        provinciasRepository.deleteById(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Integer idAuth = Integer.parseInt(auth.getCredentials().toString());
+        Usuario usuarioActual= usuariosRepository.findUsuarioById(idAuth);
+
+        if (usuarioActual.getRol().equals("ADMIN")) {
+            provinciasRepository.deleteById(id);
+        }else{
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Solo los administradores tienen permisos para eliminar");
+        }
+       
     }
 }
